# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Power Vital API",
    "version": "1.0.0",
    "description": "Power Vital e-commerce + MLM platform backend.\n\n## Languages\nAll user-facing strings live in three locales: **TR** (primary), **RU**, **KG**.\nThe active locale is negotiated via the `Accept-Language` header and a `lang`\nquery param. AI-generated translations fill missing locale fields automatically.\n\n## Auth\nMutating endpoints require a `Bearer` JWT in the `Authorization` header.\nAdmin-only routes additionally check the `role` claim.\n\n## Rate limiting\nAll `/api/*` routes are rate-limited to 150 req/min per IP. Burst over\nthe limit returns 429 with `Retry-After`.\n\n## Errors\nAll errors return `{ error: string, issues?: ValidationIssue[] }` where\n`ValidationIssue` has the shape `{ path, message, code }`."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://powervital.kg",
      "description": "Production (Kyrgyzstan)"
    },
    {
      "url": "https://powervital.org",
      "description": "Production (intl alias)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from POST /api/v1/auth/login"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message",
                "code"
              ]
            }
          }
        },
        "required": [
          "error"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 128
          },
          "sponsorId": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ],
        "description": "Self-service customer registration. Sponsor chain is read from the sponsorId when present."
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ],
        "description": "Exchange email + password for a 24h JWT. Returns the token plus a public-safe user profile."
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "minLength": 1
          },
          "newPassword": {
            "type": "string",
            "minLength": 6,
            "maxLength": 128
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "description": "Authenticated user changes their own password. Requires the current password as proof."
      },
...
    },
...
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Register a new customer account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Exchange credentials for a short-lived access token (15m) + long-lived refresh token (7d). The refresh token is also set as an HttpOnly cookie. Replay of an already-rotated refresh token revokes the entire family.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Returns accessToken + refreshToken + user profile + TTLs.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string",
                      "description": "Legacy alias for accessToken"
                    },
                    "expiresIn": {
                      "type": "number",
                      "description": "Access token TTL in seconds"
                    },
                    "refreshExpiresIn": {
                      "type": "number",
                      "description": "Refresh token TTL in seconds"
                    },
                    "tokenType": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "email",
                        "role"
                      ],
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "token",
                    "expiresIn",
                    "refreshExpiresIn",
                    "tokenType",
                    "user"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Too many login attempts (5/15min per IP)"
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Rotate the refresh token. The new refresh token is returned and (for browser clients) set as an HttpOnly cookie. The old token is marked revoked. If the supplied token has already been used, the entire family is revoked and 401 is returned. The refresh token may be sent either in the request body (refreshToken field) or via the HttpOnly pv_refresh cookie.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New access + refresh tokens issued"
          },
          "401": {
            "description": "Invalid, expired, or replayed refresh token"
          }
        }
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "Return the authenticated user profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user"
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/api/v1/auth/network": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "Return the distributor downline tree (max depth 8)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Network tree"
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/api/v1/auth/change-password": {
      "put": {
        "tags": [
          "Auth"
        ],
        "description": "Change the authenticated user's password",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Wrong current password"
          }
        }
      }
    },
...
    "/api/v1/admin/logs": {
      "get": {
        "tags": [
          "Admin — Logs"
        ],
        "description": "Tail the most recent lines of the backend log file. Filters: ?level=info|warn|error|debug, ?q=text, ?limit=200, ?maxBytes=256K, ?since=ms. Returns parsed JSON lines ordered chronologically (oldest first).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Parsed log lines"
          },
          "503": {
            "description": "Log file not configured"
          }
        }
      }
    },
...
    "/api/v1/reviews/admin/{id}": {
      "delete": {
        "tags": [
          "Product Reviews"
        ],
        "description": "Delete a review (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
...
  },
...
}